+++ src/SFA.Apprenticeships.Web.Manage/Controllers/VacancyController.cs
@@ -32,7 +32,7 @@
         // GET: Vacancy
         [HttpGet]
         [OutputCache(Duration = 0, NoStore = true, VaryByParam = "none")]
-        public ActionResult Review(long vacancyReferenceNumber)
+        public ActionResult Review(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.ReserveVacancyForQA(vacancyReferenceNumber);
             var vacancyViewModel = response.ViewModel;
@@ -73,7 +73,7 @@
 
         [HttpGet]
         [OutputCache(Duration = 0, NoStore = true, VaryByParam = "none")]
-        public ActionResult BasicDetails(long vacancyReferenceNumber)
+        public ActionResult BasicDetails(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.GetBasicDetails(vacancyReferenceNumber);
 
@@ -121,7 +121,7 @@
 
         [HttpGet]
         [OutputCache(Duration = 0, NoStore = true, VaryByParam = "none")]
-        public ActionResult TrainingDetails(long vacancyReferenceNumber)
+        public ActionResult TrainingDetails(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.GetTrainingDetails(vacancyReferenceNumber);
 
@@ -191,7 +191,7 @@
 
         [HttpGet]
         [OutputCache(Duration = 0, NoStore = true, VaryByParam = "none")]
-        public ActionResult Summary(long vacancyReferenceNumber)
+        public ActionResult Summary(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.GetVacancySummaryViewModel(vacancyReferenceNumber);
 
@@ -238,7 +238,7 @@
 
         [HttpGet]
         [OutputCache(Duration = 0, NoStore = true, VaryByParam = "none")]
-        public ActionResult RequirementsAndProspects(long vacancyReferenceNumber)
+        public ActionResult RequirementsAndProspects(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.GetVacancyRequirementsProspectsViewModel(vacancyReferenceNumber);
 
@@ -284,7 +284,7 @@
 
         [HttpGet]
         [OutputCache(Duration = 0, NoStore = true, VaryByParam = "none")]
-        public ActionResult Questions(long vacancyReferenceNumber)
+        public ActionResult Questions(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.GetVacancyQuestionsViewModel(vacancyReferenceNumber);
 
@@ -331,7 +331,7 @@
 
         [MultipleFormActionsButton(SubmitButtonActionName = "VacancyQAAction")]
         [HttpPost]
-        public ActionResult Approve(long vacancyReferenceNumber)
+        public ActionResult Approve(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.ApproveVacancy(vacancyReferenceNumber);
 
@@ -349,7 +349,7 @@
 
         [MultipleFormActionsButton(SubmitButtonActionName = "VacancyQAAction")]
         [HttpPost]
-        public ActionResult Reject(long vacancyReferenceNumber)
+        public ActionResult Reject(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.RejectVacancy(vacancyReferenceNumber);
 
@@ -366,7 +366,7 @@
         }
 
         [HttpGet]
-        public ActionResult EmployerInformation(long vacancyReferenceNumber, bool? useEmployerLocation)
+        public ActionResult EmployerInformation(int vacancyReferenceNumber, bool? useEmployerLocation)
         {
             var response = _vacancyMediator.GetEmployerInformation(vacancyReferenceNumber, useEmployerLocation);
 
@@ -411,7 +411,7 @@
         }
 
         [HttpGet]
-        public ActionResult Locations(long vacancyReferenceNumber)
+        public ActionResult Locations(int vacancyReferenceNumber)
         {
             var response = _vacancyMediator.GetLocationAddressesViewModel(vacancyReferenceNumber);
