+++ src/SFA.Apprenticeships.Web.Recruit.UnitTests/Mediators/VacancyPosting/CloneVacancyTests.cs
@@ -12,7 +12,7 @@
         [Test]
         public void ShouldReturnVacancyInIncorrectStateErrorCodeIfVacancyIsInReferredStatus()
         {
-            long vacancyReferenceNumber = 1;
+            int vacancyReferenceNumber = 1;
 
             VacancyPostingProvider.Setup(p => p.GetVacancy(vacancyReferenceNumber)).Returns(new VacancyViewModel { Status = VacancyStatus.RejectedByQA});
             var mediator = GetMediator();
