+++ src/SFA.Apprenticeships.Web.Recruit/Controllers/VacancyPostingController.cs
@@ -197,7 +197,7 @@
         }
 
         [HttpGet]
-        public ActionResult ReviewCreateVacancy(long vacancyReferenceNumber, bool? comeFromPreview)
+        public ActionResult ReviewCreateVacancy(int vacancyReferenceNumber, bool? comeFromPreview)
         {
             var response = _vacancyPostingMediator.GetNewVacancyViewModel(vacancyReferenceNumber, true, comeFromPreview);
             var viewModel = response.ViewModel;
@@ -286,7 +286,7 @@
         #region Training Details
 
         [HttpGet]
-        public ActionResult TrainingDetails(long vacancyReferenceNumber)
+        public ActionResult TrainingDetails(int vacancyReferenceNumber)
         {
             var response = _vacancyPostingMediator.GetTrainingDetailsViewModel(vacancyReferenceNumber, false, false);
             var viewModel = response.ViewModel;
@@ -316,7 +316,7 @@
         }
 
         [HttpGet]
-        public ActionResult ReviewTrainingDetails(long vacancyReferenceNumber, bool? comeFromPreview)
+        public ActionResult ReviewTrainingDetails(int vacancyReferenceNumber, bool? comeFromPreview)
         {
             var response = _vacancyPostingMediator.GetTrainingDetailsViewModel(vacancyReferenceNumber, true, comeFromPreview);
             var viewModel = response.ViewModel;
@@ -405,7 +405,7 @@
         #region Vacancy Details
 
         [HttpGet]
-        public ActionResult VacancySummary(long vacancyReferenceNumber)
+        public ActionResult VacancySummary(int vacancyReferenceNumber)
         {
             var response = _vacancyPostingMediator.GetVacancySummaryViewModel(vacancyReferenceNumber, false, false);
             var viewModel = response.ViewModel;
@@ -421,7 +421,7 @@
         }
 
         [HttpGet]
-        public ActionResult ReviewVacancySummary(long vacancyReferenceNumber, bool? comeFromPreview)
+        public ActionResult ReviewVacancySummary(int vacancyReferenceNumber, bool? comeFromPreview)
         {
             var response = _vacancyPostingMediator.GetVacancySummaryViewModel(vacancyReferenceNumber, true, comeFromPreview);
             var viewModel = response.ViewModel;
@@ -503,7 +503,7 @@
         #region Requirements and Prospects
 
         [HttpGet]
-        public ActionResult VacancyRequirementsProspects(long vacancyReferenceNumber, bool? comeFromPreview)
+        public ActionResult VacancyRequirementsProspects(int vacancyReferenceNumber, bool? comeFromPreview)
         {
             var response = _vacancyPostingMediator.GetVacancyRequirementsProspectsViewModel(vacancyReferenceNumber, false, comeFromPreview);
             var viewModel = response.ViewModel;
@@ -519,7 +519,7 @@
         }
 
         [HttpGet]
-        public ActionResult ReviewVacancyRequirementsProspects(long vacancyReferenceNumber)
+        public ActionResult ReviewVacancyRequirementsProspects(int vacancyReferenceNumber)
         {
             var response = _vacancyPostingMediator.GetVacancyRequirementsProspectsViewModel(vacancyReferenceNumber, true, true);
             var viewModel = response.ViewModel;
@@ -598,7 +598,7 @@
         #region Vacancy Questions
 
         [HttpGet]
-        public ActionResult VacancyQuestions(long vacancyReferenceNumber, bool? comeFromPreview)
+        public ActionResult VacancyQuestions(int vacancyReferenceNumber, bool? comeFromPreview)
         {
             var response = _vacancyPostingMediator.GetVacancyQuestionsViewModel(vacancyReferenceNumber, false, comeFromPreview);
             var viewModel = response.ViewModel;
@@ -613,7 +613,7 @@
             }
         }
 
-        public ActionResult ReviewVacancyQuestions(long vacancyReferenceNumber, bool? comeFromPreview)
+        public ActionResult ReviewVacancyQuestions(int vacancyReferenceNumber, bool? comeFromPreview)
         {
             var response = _vacancyPostingMediator.GetVacancyQuestionsViewModel(vacancyReferenceNumber, true, comeFromPreview);
             var viewModel = response.ViewModel;
@@ -682,7 +682,7 @@
 
         [HttpGet]
         [OutputCache(Duration = 0)]
-        public ActionResult PreviewVacancy(long vacancyReferenceNumber)
+        public ActionResult PreviewVacancy(int vacancyReferenceNumber)
         {
             var response = _vacancyPostingMediator.GetPreviewVacancyViewModel(vacancyReferenceNumber);
 
@@ -720,7 +720,7 @@
         }
         
         [HttpPost]
-        public ActionResult SubmitVacancy(long vacancyReferenceNumber, bool resubmitoption)
+        public ActionResult SubmitVacancy(int vacancyReferenceNumber, bool resubmitoption)
         {
             var response = _vacancyPostingMediator.SubmitVacancy(vacancyReferenceNumber, resubmitoption);
             var vacancyViewModel = response.ViewModel;
@@ -750,7 +750,7 @@
         }
 
         [HttpGet]
-        public ActionResult VacancySubmitted(long vacancyReferenceNumber, bool resubmitted)
+        public ActionResult VacancySubmitted(int vacancyReferenceNumber, bool resubmitted)
         {
             var response = _vacancyPostingMediator.GetSubmittedVacancyViewModel(vacancyReferenceNumber, resubmitted);
             var viewModel = response.ViewModel;
@@ -964,7 +964,7 @@
         }
 
         [HttpGet]
-        public ActionResult ManageDates(long vacancyReferenceNumber)
+        public ActionResult ManageDates(int vacancyReferenceNumber)
         {
             var response = _vacancyPostingMediator.GetVacancyDatesViewModel(vacancyReferenceNumber);
